requirements.txt
python-telegram-bot==13.3
requests==2.25.1
lxml==4.6.2